+++ docs/@views/api/samples/createIndicator-stack/index.js
@@ -1,11 +1,11 @@
 import { init } from 'klinecharts'
 
 const chart = init('createIndicator-stack-chart')
-chart.createIndicator('MA', false, { id: 'candle_pane' })
-chart.createIndicator('EMA', true, { id: 'candle_pane' })
+chart.createIndicator({ name: 'MA', paneId: 'candle_pane' })
+chart.createIndicator({ name: 'EMA', paneId: 'candle_pane' }, { isStack: true })
 
 const id = chart.createIndicator('BOLL')
-chart.createIndicator('EMA', true, { id })
+chart.createIndicator({ name: 'EMA', paneId: id }, { isStack: true })
 
 chart.setSymbol({ ticker: 'TestSymbol' })
 chart.setPeriod({ span: 1, type: 'day' })