+++ docs/@views/api/samples/getIndicators/index.js
@@ -1,7 +1,7 @@
 import { init } from 'klinecharts'
 
 const chart = init('getIndicators-chart')
-chart.createIndicator('MA', false, { id: 'candle_pane' })
+chart.createIndicator({ name: 'MA', paneId: 'candle_pane' })
 chart.createIndicator('VOL')
 
 chart.setSymbol({ ticker: 'TestSymbol' })