+++ docs/@views/api/samples/overrideIndicator-basic/index.js
@@ -1,7 +1,7 @@
 import { init } from 'klinecharts'
 
 const chart = init('overrideIndicator-basic-chart')
-chart.createIndicator('MA', false, { id: 'candle_pane' })
+chart.createIndicator({ name: 'MA', paneId: 'candle_pane' })
 chart.createIndicator('MA')
 chart.overrideIndicator({
   name: 'MA',