+++ docs/@views/api/samples/setPaneOptions-axis-extension/index.js
@@ -3,14 +3,12 @@ import { init } from 'klinecharts'
 const chart = init('setPaneOptions-axis-extension-chart')
 chart.createIndicator('MACD')
 
-chart.setPaneOptions({
-  id: 'candle_pane',
-  axis: {
-    createTicks: ({ defaultTicks }) => {
-      return defaultTicks.map(
-        ({ coord, value, text }) => ({ coord, value, text: `$${text}` })
-      )
-    }
+chart.overrideYAxis({
+  paneId: 'candle_pane',
+  createTicks: ({ defaultTicks }) => {
+    return defaultTicks.map(
+      ({ coord, value, text }) => ({ coord, value, text: `$${text}` })
+    )
   }
 })