+++ src/index.ts
@@ -47,6 +47,7 @@ import { calcTextWidth } from './common/utils/canvas'
 import type { ActionType } from './common/Action'
 import type { IndicatorSeries } from './component/Indicator'
 import type { OverlayMode } from './component/Overlay'
+import type { AxisOptions, PaneOptions } from './pane/types'
 
 import type { FormatDateType, Options, ZoomAnchor } from './Options'
 import ChartImp, { type Chart, type DomPosition } from './Chart'
@@ -175,5 +176,5 @@ export {
   utils,
   type LineType, type PolygonType, type TooltipShowRule, type TooltipShowType, type FeatureType, type TooltipFeaturePosition, type CandleTooltipRectPosition,
   type CandleType, type FormatDateType, type ZoomAnchor,
-  type DomPosition, type ActionType, type IndicatorSeries, type OverlayMode
+  type DomPosition, type ActionType, type IndicatorSeries, type OverlayMode, type AxisOptions, type PaneOptions
 }