+++ src/pane/types.ts
@@ -12,8 +12,6 @@
  * limitations under the License.
  */
 
-import type { AxisCreate } from '../component/Axis'
-
 // export type PaneState = 'normal' | 'maximize' | 'minimize'
 
 export interface PaneOptions {
@@ -22,7 +20,6 @@ export interface PaneOptions {
   minHeight?: number
   dragEnabled?: boolean
   order?: number
-  axis?: Partial<AxisCreate>
 }
 
 export const PANE_MIN_HEIGHT = 30
@@ -33,8 +30,4 @@ export const PaneIdConstants = {
   CANDLE: 'candle_pane',
   INDICATOR: 'indicator_pane_',
   X_AXIS: 'x_axis_pane'
-}
-
-export const YAxisIdConstants = {
-  DEFAULT: 'default'
 }