+++ src/view/CrosshairHorizontalLabelView.ts
@@ -19,9 +19,8 @@ import { isString } from '../common/utils/typeChecks'
 import { createFont } from '../common/utils/canvas'
 import { SymbolDefaultPrecisionConstants } from '../common/SymbolInfo'
 
-import type { Axis } from '../component/Axis'
+import { DEFAULT_AXIS_ID, type Axis } from '../component/Axis'
 import type YAxis from '../component/YAxis'
-import { YAxisIdConstants } from '../pane/types'
 
 import type { TextAttrs } from '../extension/figure/text'
 
@@ -71,7 +70,7 @@ export default class CrosshairHorizontalLabelView<C extends Axis = YAxis> extend
     const value = axis.convertFromPixel(crosshair.y!)
     let precision = 0
     let shouldFormatBigNumber = false
-    if (yAxis.isInCandle() && yAxis.id === YAxisIdConstants.DEFAULT) {
+    if (yAxis.isInCandle() && yAxis.id === DEFAULT_AXIS_ID) {
       precision = chartStore.getSymbol()?.pricePrecision ?? SymbolDefaultPrecisionConstants.PRICE
     } else {
       const indicators = chartStore.getIndicatorsByPaneId(crosshair.paneId!).filter(indicator => indicator.yAxisId === yAxis.id)